requirements-dev.txt
Faker==1.0.5
Sphinx==2.0.1
aiohttp-theme==0.1.5
aioresponses==0.6.0
asynctest==0.12.4
bpython
flake8==3.7.7
py-spy==0.1.11
pydocstyle==3.0.0
pyinstrument==3.0.3
pylint==2.4.1
pytest-asyncio==0.10.0
pytest-bdd==3.1.0
pytest-cov==2.6.1
pytest-mock==1.10.4
pytest-rerunfailures==7.0
pytest-socket==0.3.3
pytest-xdist==1.28.0
pytest==4.4.1
q==2.6
recommonmark==0.5.0
sphinx-bootstrap-theme==0.7.1
sphinx-click==2.0.1
tox==3.9.0
twine==1.13.0
wheel==0.33.1
yapf==0.27.0
